requirements.txt
flake8==3.9.2
isort==5.9.3
mccabe==0.6.1
pycodestyle==2.7.0
pyflakes==2.3.1
python-dotenv==0.19.1
